src/backend/requirements.txt
# CodeClinic Backend Dependencies
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP requests and ZAP integration
httpx==0.25.2
requests==2.31.0

# Data validation and processing
pydantic==2.5.0
pydantic-settings==2.1.0

# URL validation
validators==0.22.0

# Environment variables
python-dotenv==1.0.0

# Parallel scanning and coordination
redis==5.0.1

# Performance and concurrency (concurrent.futures is built-in for Python 3.8+)

# ZAP Python API client
python-owasp-zap-v2.4==0.0.20

# Optional: For future AI integration
# google-generativeai==0.3.2
